[GOORM-65]-refactor: 가족 질문 패키지명 수정
