+++ src/main/java/goorm/kgu/familynote/domain/family/question/domain/FamilyQuestionRepository.java
@@ -1,4 +1,4 @@
-package goorm.kgu.familynote.domain.family.familyQuestion.domain;
+package goorm.kgu.familynote.domain.family.question.domain;
 
 
 import java.util.List;
