+++ src/main/java/goorm/kgu/familynote/domain/family/question/presentation/response/FamilyQuestionPageResponse.java
@@ -1,4 +1,4 @@
-package goorm.kgu.familynote.domain.family.familyQuestion.presentation.response;
+package goorm.kgu.familynote.domain.family.question.presentation.response;
 
 import goorm.kgu.familynote.common.response.PageableResponse;
 import io.swagger.v3.oas.annotations.media.Schema;
